pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.3.1",
    "init>=0.1.0",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
